Correct behavior of drawString() in headless mode
Symptoms:

When operating in headless mode ( i.e. java - Djava.awt.headless=true-jar
MyConverter.jar ), ImageJ is unable to perform a successful call to
ij.processor.ImageProcessor.drawString() because drawString() calls the
setupFrame() function, which in turn attempts to instanciate a
java.awt.Frame object.

This attempt results in a HeadlessException error being thrown by the
java.awt.Frame constructor (there isn't much sense for Java to create a
frame if a display server is not present, so it throws an exception error).
The mistake here is that functions that do image manipulation, like
drawString(), and do not require the ImageJ GUI, should not call
setupFrame() if Java is running in headless mode.

Solution:

The only reason drawString() calls setupFrame() is to create valid Font
and FontMetrics objects, and the only reason setupFrame() creates a Frame
object is to get a Graphics object that is used to create the FontMetrics
object. If it were possible to get the Graphics object from a different
source, not java.awt.Frame, then there would be no need to create the
Frame.

And thus, I have modified the setupFrame() function to use a BufferedImage
object, only in case the java.awt.headless property is true.

[dscho: minimal style fixups]

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>