Fix bug in ROI splitting
There was a bug with the last extracted Roi when splitting with the ROI
Manager.

It turns out that it is just an off-by-one bug: the split loop in
parsePath() will increment "count" before seeing a SEG_MOVETO (which
triggers the adding of the current Roi), while the variable "count"
is not incremented when "done" becomes true, and the last Roi is added.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>