Make window icon overrideable with -icon=<path>
We need this in Fiji to prevent the ImageJ window from popping up before
the window icon is set in StartupMacros.

This does not work on MacOSX, as its Java does not allow overriding the
window icon, except by the command line option "-Xdock:icon=<path>"
(which we use in Fiji).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>