Added the method int getPixelInterpolated( double, double ) to
ij.ImageProcessor

and implemented it in

ij.ByteProcessor
ij.ShortProcessor
ij.FloatProcessor
ij.ColorProcessor

The method returns an integer conformous to int getPixel( int, int ) linearly
interpolated fo real image coordinates.  I expected this from the method
getInterpolatedPixel( double, double ) but that returns an interpolated double
with the exception of ColorProcessor where it returns a non-interpolated double.The mentioned functionality is required for implementing a geometric mapping
with putPixel( int, int, getPixelInterpolated( double, double ) ) which is valid
for all ImageProcessors.