Replace the SocketListener by an RMI-based listener
The problem with the SocketListener is that on a multi-user machine,
other users can connect to your ImageJ instance, sending and running
macros and javascripts without you noticing.

Even worse: if you do not have a firewall, this will be possible from
other computers as well (and guessing the port is easy).

Instead, use an RMI based listener, as RMI has solved all the security
problems for us already.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>