Replace the SocketListener by an RMI-based listener
The problem with the SocketListener is that on a multi-user machine,
other users can connect to your ImageJ instance, sending and running
macros and javascripts without you noticing.

Even worse: if you do not have a firewall, this will be possible from
other computers -- without authentication -- as well (and guessing the
port is easy).

Instead, use an RMI (Remote Method Invocation, a mechanism to access
remote objects) based listener, as RMI has solved all the security
problems for us already.

The serialized stub (which is what is needed to connect to a
remote object) is written to the temporary directory and made private,
to be picked up by other calls to ImageJ by the same user, on the same
machine.

Prior to Java5, the stubs for remote objects are not generated on the
fly, but have to be generated using the program 'rmic'.  And as ImageJ
tries to be backwards-compatible up to Java2, we have to teach Ant to
generate the stub and include it in ij.jar, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>