CommandFinder: don't completely clear the list to update it
Instead of completely clearing and repopulating the list of
matching commands, step through the original list and the new
list making minimal additions and removals.  Repopulating the
list from empty means that the scrollbar changes in a difficult
way for the user to understand, and it's probably faster to
update with this change.