Fix Unicode issues with Tiff opening
The default encoding of strings in Java is not ISO-8859-1, but many
.tif images (e.g. ImageJ's Fly Brain sample image) contain information
in that encoding.

Fix reading this information by falling back to decode from ISO-8859-1
when UTF-8 failed.

Care has to be taken, too, when using the Properties class' load()
method, as it expects the input in ISO-8859-1 format (for backwards
compatibility), but it can interpret Unicode characters escaped in
that funny \uABCD format.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>