Command Finder: pretend that the prompt is index -1 of the list
When pressing cursor down in the prompt, it selects the first item in the
list, which is in line with the list being just below the prompt.  At the
same time, pressing cursor down, the last item in the list is selected.

It is now logical to interpret cursor up when the first item is selected
as meaning "go back to the prompt", and likewise cursor down on the last
item.

Lastly, page down in the prompt will select the last visible item in the
list, but page up in the first page of the list will first go to the
first list item and only then to the prompt.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>