Put back public APIs that have never been marked deprecated
A comment in the source code that a method is obsolete can be missed
easily. Code that is marked @deprecated is less easy to miss, so let's
put back the methods for a certain grace period, to give plugin
developers a chance to modify their source code accordingly.

Suggested by Stephan Saalfeld.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>