Revert: Fix LZW symbol table overflow
This was sent upstream.

Some .lsm files seem to miss the extra 0 bit when the EOI code is _just_ after
increasing the bit size of the codes to be read. Work around this issue (the
LZW compression is invalid by the TIFF specification) by limiting the number
of bytes.

Noticed by Nagananda Gurudev.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>