Make instances dependent on DISPLAY
Usually, with multi-user setups different users have different login
names. In some cases, however, this is not the case. Still, when such
users log in remotely via SSH, their DISPLAY is set differently. So
let's make the name of the stub dependent on the DISPLAY environment
variable in addition to the user name.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>