Work around MacOSX bug in FileDialog (save)
MacOSX' FileDialog automatically strips the extension. That would not be
bad if the FileDialog would keep the extension internally, but it does
not. So let's give it a fake extension it can strip all it likes.

The previous solution looked on all platforms whether the return value
still has an extension and added one in the negative case. This solution
has the following shortcomings:

- the save dialog looks at the wrong file when checking whether to ask the
  user "Overwrite?"

- if the filename contains another dot in addition to the dot separating
  name from extension, ImageJ fails to add back the extension

- if the user chose a file name without an extension at all, ImageJ will
  try to be clever and add the extension back

Of course, as soon as Apple decides to fix that bug in their FileDialog,
we have to work around the occasional need for a workaround.

This issue was reported by Helena Jambor.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>