Standardize groupId to match Maven convention.
This change will help enable eventual distribution of ImageJ through the
Maven central repository. The requirement for groupId is that it
correspond to Java packaging convention of reverse domain name. Since we
registered imagej.net and plan for that domain to serve as the central
ImageJ resource, the groupId needs to be net.imagej.