Manually fix up parent version
The previous run of the Jenkins job labeled 'Synchronize and deploy IJ1'
looked in the old spot when determining the latest pom-scijava version.
Since we deploy both pom-scijava and ij to OSS Sonatype now, the script
needs to look in the new spot, though.

The script has been fixed, now let's fix the deployment of 1.48r
manually. (The downside is that this once, the deployed 1.48r will
look *slightly* different than the tagged one: the parent project
version. On the other hand, the commit name is recorded in the
manifest, so it is not all that bad.)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>