Without this line, an Exception window opens, when you call
run("Create selection");

from IJ macro. The exception is a NullPointerException in Line 896/897 (in ShapeRoi.java)

segType = pIter.currentSegment(coords);

With this line, the function does the same thing in macro, like when using the "Create Selection" menu from ImageJ, it creates no selection. I guess whoever wrote a macro where this could happen, had to write an empty-ROI-check anyhow, as the exception does not stop the macro. So preexisting macros should not be affected.