Require Java 6
This project now uses Java-6-specific source code.

For example, the iterator() method of ij.gui.PointRoi creates an inner
class extending Iterator<Point> which decorates the methods hasNext(),
next() and remove() with @Override. Using the @Override annotation in
this way on interface methods (as opposed to abstract class methods)
is not allowed in Java 1.5, but only starting with Java 6.

Without this change, the project no longer builds in the Eclipse IDE
(and maybe other IDEs).