Fix scroll wheel zoom in image stacks
When zooming in image stacks (i.e. StackWindow instances), zooming with "+" and "-" keyboard shortcuts zooms in on the pixel the cursor is hovering over, as indented. However, before this fix, zooming with CTRL or SHIFT + mouse wheel only zoomed in on a fixed location. It is worth mentioning that on non-stacked images (ImageWindow instances), this was not an issue.