Enable compilation of module-info.java
Adapted from
https://maven.apache.org/plugins/maven-compiler-plugin/examples/module-info.html
as suggested by @imagejan.

This is necessary due to ImageJ 1.x adding a module-info.java (part of
the Java 9 jigsaw module ecosystem).

The conventional wisdom would be to compile everything once with J9, and
then everything but module-info.java with J8. Unfortunately I was not
able to actually compile with J9 due to the MacAdapter.java, presumably
not interacting well with modularization.

As a work-around we compile *only* module-info.java with J11, and then
everything else with J8. I assume this leads to a useless module-info,
but it's not clear why it was added in the first place.

Note that this project now requires Java 9+ to compile, but can still
run on J6+. We select J11 due to its LTS release status.