Exclude MacAdapter from module build
MacAdapter imports mac-specific internal packages. We use the Orange
Extensions to provide stubs for these packages, which is fine in Java 8
or earlier.

However when using Java 9+ to build modules, to satisfy MacAdapter's
package requirements we need to add an orange.extensions requirement to
the module-info.java. This is a file-based dependency using Maven's
automodule functionality.

MacAdapter in IJ1 is not meant to be distributed on non-Macs. To avoid
the need to edit the module-info.java and introduce a file-based
dependency (with no guarantee of it being present in IJ1) we instead
exclude MacAdapter from the module-building process.

MacAdapter and the orange extension stubs will still be pulled in via
Maven.