Updated Fit Rectangle to always be oriented from top to bottom
The new rotated rectangle is now generated after checking that: a) The first control point is on the shorter side. This corrects erratic orientations when the original ROI is rectangular.
b) The first control point is at the top.

This ensures that the rectangle will run from top to bottom of the image, startarting from a short side, resulting in a more consistent orientation after cropping (top always to the left side).

This was discussed in this forum post: https://forum.image.sc/t/how-to-control-rotation-of-a-rotated-rectangle-selection-during-image-duplication/55515